AlgoRelationType add Assembler and EntityConverter
refactoring of validation

+++ org.planqk.atlas.core/src/main/java/org/planqk/atlas/core/services/AlgoRelationTypeServiceImpl.java
@@ -16,6 +16,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -35,7 +37,7 @@ public class AlgoRelationTypeServiceImpl implements AlgoRelationTypeService {
 	@Override
 	public AlgoRelationType update(UUID id, AlgoRelationType algoRelationType) throws NotFoundException {
 		// Check for type in database
-		Optional<AlgoRelationType> typeOpt = findById(id);
+		Optional<AlgoRelationType> typeOpt = findOptionalById(id);
 		// If Type exists
 		if (typeOpt.isPresent()) {
 			// Update fields
@@ -49,27 +51,44 @@ public class AlgoRelationTypeServiceImpl implements AlgoRelationTypeService {
 	}
 
 	@Override
-	public void delete(UUID id) throws SqlConsistencyException {
+	public void delete(UUID id) throws SqlConsistencyException, NotFoundException {
 		if (algorithmRelationRepository.countRelationsUsingRelationType(id) > 0) {
 			LOG.info("Trying to delete algoRelationType that is used in at least 1 algorithmRelation.");
-			throw new SqlConsistencyException("Cannot delete algoRelationType, since it is used by existing algorithmRelations!");
+			throw new SqlConsistencyException("Cannot delete algoRelationType since it is used by existing algorithmRelations.");
+		}
+		if (repo.findById(id).isEmpty()) {
+			LOG.info("Trying to delete algoRelationType which does not exist.");
+			throw new NotFoundException("Cannot delete algoRelationTypesince it could not be found.");
 		}
 		repo.deleteById(id);
 	}
 
 	@Override
-	public Optional<AlgoRelationType> findById(UUID id) {
-		return Objects.isNull(id) ? Optional.empty() : repo.findById(id);
+	public AlgoRelationType findById(UUID id) throws NotFoundException {
+		Optional<AlgoRelationType> algoRelationTypeOpt = findOptionalById(id);
+		if (algoRelationTypeOpt.isEmpty()) {
+			throw new NotFoundException("The AlgoRelationType could not be found.");
+		}
+		return algoRelationTypeOpt.get();
 	}
 
 	@Override
-	public Optional<List<AlgoRelationType>> findByName(String name) {
-		return Objects.isNull(name) ? Optional.empty() : repo.findByName(name);
+	public List<AlgoRelationType> findByName(String name) throws NotFoundException {
+		Optional<List<AlgoRelationType>> algoRelationTypes = repo.findByName(name);
+		if (algoRelationTypes.isEmpty()) {
+			throw new NotFoundException("No AlgoRelationType found to match name '" + name + "'");
+		}
+		return algoRelationTypes.get();
 	}
 
 	@Override
 	public Page<AlgoRelationType> findAll(Pageable pageable) {
 		return repo.findAll(pageable);
+	}
+
+	@Override
+	public Optional<AlgoRelationType> findOptionalById(UUID id) {
+		return repo.findById(id);
 	}
 
 }
